Implement universal dlopen-style API
This adds `external_library_universal!`, a new entrypoint to the crate.
It's very similar to `external_library!` but the generated bindings are
universal for both dlopen and non-dlopen builds.

This fixes #12. Of course, this does not address the concern about
structure size - but this is why this feature is optional.